+++ Models/Review/ReviewDto.cs
@@ -11,9 +11,9 @@ namespace Models.Review;
 /// </summary>
 public class ReviewDto
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Comment { get; set; } = string.Empty;
     public int Rating { get; set; }
     public string UserName { get; set; } = string.Empty;
-    public int MovieId { get; set; }
+    public long MovieId { get; set; }
 }
